Issue #561: Binary DATE_PART result
Fix the result type of the binary DATE_PART implementation
to be BIGINT.
Fix the test results for epoch extraction.

+++ extension/icu/icu-datepart.cpp
@@ -284,7 +284,7 @@ struct ICUDatePart {
 	}
 
 	static ScalarFunction GetBinaryPartCodeFunction(const string &name) {
-		return ScalarFunction(name, {LogicalType::VARCHAR, LogicalType::TIMESTAMP_TZ}, LogicalType::INTEGER,
+		return ScalarFunction(name, {LogicalType::VARCHAR, LogicalType::TIMESTAMP_TZ}, LogicalType::BIGINT,
 		                      BinaryFunction, false, BindDatePart);
 	}
